+++ lib/python_version.sh
@@ -8,13 +8,13 @@ LATEST_PYTHON_3_9="3.9.25"
 LATEST_PYTHON_3_10="3.10.19"
 LATEST_PYTHON_3_11="3.11.14"
 LATEST_PYTHON_3_12="3.12.12"
-LATEST_PYTHON_3_13="3.13.9"
-LATEST_PYTHON_3_14="3.14.0"
+LATEST_PYTHON_3_13="3.13.10"
+LATEST_PYTHON_3_14="3.14.1"
 
 OLDEST_SUPPORTED_PYTHON_3_MINOR_VERSION=9
 NEWEST_SUPPORTED_PYTHON_3_MINOR_VERSION=14
 
-DEFAULT_PYTHON_FULL_VERSION="${LATEST_PYTHON_3_13}"
+DEFAULT_PYTHON_FULL_VERSION="${LATEST_PYTHON_3_14}"
 DEFAULT_PYTHON_MAJOR_VERSION="${DEFAULT_PYTHON_FULL_VERSION%.*}"
 
 # Integer with no redundant leading zeros.
@@ -692,12 +692,29 @@ function python_version::warn_if_deprecated_major_version() {
 		output::warning <<-EOF
 			Warning: Support for Python 3.9 is ending soon!
 
-			Python 3.9 will reach its upstream end-of-life in October 2025,
-			at which point it will no longer receive security updates:
+			Python 3.9 reached its upstream end-of-life on 31st October 2025,
+			and so no longer receives security updates:
 			https://devguide.python.org/versions/#supported-versions
 
 			As such, support for Python 3.9 will be removed from this
 			buildpack on 7th January 2026.
+
+			Upgrade to a newer Python version as soon as possible, by
+			changing the version in your ${version_origin} file.
+
+			For more information, see:
+			https://devcenter.heroku.com/articles/python-support#supported-python-versions
+		EOF
+	elif [[ "${requested_major_version}" == "3.10" ]]; then
+		output::warning <<-EOF
+			Warning: Support for Python 3.10 is deprecated!
+
+			Python 3.10 will reach its upstream end-of-life in October 2026,
+			at which point it will no longer receive security updates:
+			https://devguide.python.org/versions/#supported-versions
+
+			As such, support for Python 3.10 will be removed from this
+			buildpack on 6th January 2027.
 
 			Upgrade to a newer Python version as soon as possible, by
 			changing the version in your ${version_origin} file.